+++ app/src/main/kotlin/ac/mdiq/podcini/storage/database/LogsAndStats.kt
@@ -56,7 +56,7 @@ object LogsAndStats {
                 feedTotalTime += m.duration
                 if (m.lastPlayedTime in timeFilterFrom..<timeFilterTo) {
                     if (includeMarkedAsPlayed) {
-                        if ((m.playbackCompletionTime > 0 && m.playedDuration > 0) || m.episodeOrFetch()?.playState == PlayState.PLAYED.code || m.position > 0) {
+                        if ((m.playbackCompletionTime > 0 && m.playedDuration > 0) || (m.episodeOrFetch()?.playState?:-10) > PlayState.SKIPPED.code || m.position > 0) {
                             episodesStarted += 1
                             feedPlayedTime += m.duration
                         }